moved new kwm dialog to dialogs folder
